Issue #43: Standalone module for creating shells of document definitions
Parses a given document definitions string while stubbing out dependent global functions and variables. Currently used only for validation purposes.
