+++ src/document-definitions-loader.js
@@ -8,20 +8,8 @@
  */
 exports.load = loadFromFile;
 
-/**
- * Parses the given document definitions string as JavaScript.
- *
- * @param {string} docDefinitionsString The document definitions as a string
- * @param {string} [originalFilename] The optional name/path of the file from which the document definitions were read. Will be used in
- *                                    stack traces.
- *
- * @returns The document definitions as a JavaScript object or function
- */
-exports.parseDocDefinitions = parseDocDefinitions;
-
 var fs = require('fs');
 var path = require('path');
-var vm = require('vm');
 var fileFragmentLoader = require('./file-fragment-loader.js');
 
 function loadFromFile(docDefinitionsFile) {
@@ -42,31 +30,4 @@ function loadFromFile(docDefinitionsFile) {
 
   // Automatically replace instances of the "importDocumentDefinitionFragment" macro with the contents of the file that is specified by each
   return fileFragmentLoader.load(docDefinitionsDir, 'importDocumentDefinitionFragment', docDefinitions);
-}
-
-function parseDocDefinitions(docDefinitionsString, originalFilename) {
-  // Fake the various global variables and functions that are available to document definitions
-  var sandbox = {
-    doc: { },
-    oldDoc: { },
-    typeIdValidator: { },
-    simpleTypeFilter: function() { return true; },
-    isDocumentMissingOrDeleted: function() { return false; },
-    isValueNullOrUndefined: function() { return false; },
-    getEffectiveOldDoc: function() { return oldDoc; },
-    requireAccess: function() { },
-    requireRole: function() { },
-    requireUser: function() { },
-    channel: function() { },
-    access: function() { },
-    role: function() { }
-  };
-  var options = {
-    filename: originalFilename,
-    displayErrors: true
-  };
-
-  var rawDocDefinitions = vm.runInNewContext('(' + docDefinitionsString + ');', sandbox, options);
-
-  return rawDocDefinitions;
 }
